your message

+++ A1/Program.cs
@@ -52,7 +52,7 @@
                     }
                 }
             }
-            uint mitja = i_ageTotal / i_numAges;
+            float mitja = i_ageTotal / i_numAges;
             Console.WriteLine($"La mitja es de {mitja} anys");
         }
     }
